Added an example of an insert, modified tables and repositories.
Many to many relations don't work.

+++ src/main/java/controller/ActorRepository.java
@@ -3,6 +3,6 @@ package controller;
 import model.Actor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActorRepository extends JpaRepository<Actor, Integer> {
+public interface ActorRepository extends JpaRepository<Actor, Long> {
 
 }
