+++ src/main/java/controller/CountryRepository.java
@@ -3,5 +3,5 @@ package controller;
 import model.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountryRepository extends JpaRepository<Country, Integer> {
+public interface CountryRepository extends JpaRepository<Country, Long> {
 }
