+++ src/main/java/controller/LanguageRepository.java
@@ -3,5 +3,5 @@ package controller;
 import model.Language;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LanguageRepository extends JpaRepository<Language, Integer> {
+public interface LanguageRepository extends JpaRepository<Language, Long> {
 }
