+++ src/main/java/controller/RatingsRepository.java
@@ -3,5 +3,5 @@ package controller;
 import model.Ratings;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RatingsRepository extends JpaRepository<Ratings, Integer> {
+public interface RatingsRepository extends JpaRepository<Ratings, Long> {
 }
