+++ src/main/java/controller/SeriesLogRepository.java
@@ -3,5 +3,5 @@ package controller;
 import model.SeriesLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeriesLogRepository extends JpaRepository<SeriesLog, Integer> {
+public interface SeriesLogRepository extends JpaRepository<SeriesLog, Long> {
 }
