+++ src/main/java/controller/WriterRepository.java
@@ -3,5 +3,5 @@ package controller;
 import model.Writer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WriterRepository extends JpaRepository<Writer, Integer> {
+public interface WriterRepository extends JpaRepository<Writer, Long> {
 }
